backend/requirements.txt
Flask
Flask-Cors
groq
python-dotenv